rails 5.0 support
The previous version did not correctly supported rails 5. When there were includes(:..) on a relation there was
an ActiveRecord::ImmutableRelation thrown. protector was modifying the relation after loading it.

We found the https://github.com/justin-rhoades/protector/tree/prep-gem-for-upgrade branch that seems to resolve this.
In this branch Justin is removing alias_chain_method and is also resetting the relation properly with a .reset method.
This prevents the ActiveRecord::ImmutableRelation error from occuring

I have taken some of the changes of prep-gem-for-upgrade. Not all of them because they are way to many, where many of them
are connected with formating the specs and some appraises.