[renovation2022] Trying to run the specs in 2020 is impossible. Bringing them back
There are two main reasons for now

1. bundle install will get activesupport 7.0.4. But in activesupport 5 a method is removed and there is an error

An error occurred in an `after(:suite)` hook.
Failure/Error:
  silence_stream(STDOUT) do
    SimpleCov::Formatter::HTMLFormatter.new.format(SimpleCov.result)
  end

NoMethodError:
  undefined method `silence_stream' for #<RSpec::Core::AnonymousExampleGroup (no description provided)>
  Did you mean?  silence_warnings
 ./spec/spec_helpers/boot.rb:29:in `block (2 levels) in <top (required)>'

silence_stream is removed. For this we are fixing the version of activesupport that we can support to < 5

2. The second issue is with simplecov-summary.
This gem should be removed but I want to bring the gem to a version that is working as it should have worked in 2015-2016 and
make sure that tests are passing without modificiations.

Because of this I am fixing simplecov-summary to 0.0.4.

In this way I can run the specs. Otherwise an error occurs

An error occurred in an `after(:suite)` hook.
Failure/Error: SimpleCov::Formatter::SummaryFormatter.new.format(SimpleCov.result)

TypeError:
  no implicit conversion of Symbol into String
 /Users/kireto/.rvm/gems/ruby-2.7.6/gems/colored-1.2/lib/colored.rb:70:in `colorize'
 /Users/kireto/.rvm/gems/ruby-2.7.6/gems/simplecov-summary-0.0.6/lib/simplecov-summary.rb:27:in `block in format'
 /Users/kireto/.rvm/gems/ruby-2.7.6/gems/simplecov-summary-0.0.6/lib/simplecov-summary.rb:15:in `each'
 /Users/kireto/.rvm/gems/ruby-2.7.6/gems/simplecov-summary-0.0.6/lib/simplecov-summary.rb:15:in `format'
 ./spec/spec_helpers/boot.rb:33:in `block (2 levels) in <top (required)>'