Read twice, code once
Fixing the bug I generated last time I tried to fix the "undefined method" caused by "alias_method".
Read what changes are made here and applied them to us
https://github.com/rails/rails/pull/39098

This is soooooo obscure not a single mention around rails 6.1 so there was no way to know this. The issue is that by defining it this way there is no easy way to byebug into the super method and see that it is also defined like that, no way... The good thing is that we had tests and we caught it. 

We must be careful with protector it is seriously not maintained.